"Vocabulary.getTokenCount()" now returns long instead of int

+++ src/main/java/se/kth/speech/coin/tangrams/wac/data/Vocabulary.java
@@ -60,8 +60,8 @@ public final class Vocabulary {
 		return wordObservationCounts.getLong(word);
 	}
 
-	public int getTokenCount() {
-		int result = 0;
+	public long getTokenCount() {
+		long result = 0L;
 		for (final long count : wordObservationCounts.values()) {
 			result += count;
 		}
